Refactor table toolbar into separate file
