+++ src/components/ModList/index.ts
@@ -1,0 +1,3 @@
+import ModList from './ModList';
+
+export default ModList;
